requirements.txt
openai
requests
pillow
django
djangorestframework
tinydb